Added the "getStatistics" functions in the Distributions and also added the Samples classes
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/uncertwebrtools@9 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/R/ClassCenteredMoment.R
@@ -1,6 +1,6 @@
 setClass(
   Class="CenteredMoment",
-  representation=representation(values="numeric", order="numeric")
+  representation=representation(values="numeric", order="integer")
   )
 
 setMethod(
